+++ astropy/visualization/stretch.py
@@ -11,15 +11,16 @@ import numpy as np
 
 from ..extern import six
 from ..utils.misc import InheritDocstrings
-
 from .transform import BaseTransform
+
 
 __all__ = ["BaseStretch", "LinearStretch", "SqrtStretch", "PowerStretch",
            "PowerDistStretch", "SquaredStretch", "LogStretch", "AsinhStretch",
            "SinhStretch", "HistEqStretch", "ContrastBiasStretch"]
 
 
-def logn(n, x, out=None):
+def _logn(n, x, out=None):
+    """Calculate the log base n of x."""
     # We define this because numpy.lib.scimath.logn doesn't support out=
     if out is None:
         return np.log(x) / np.log(n)
@@ -29,11 +30,12 @@ def logn(n, x, out=None):
         return out
 
 
-def _prepare(values, out=None, clip=True):
+def _prepare(values, clip=True, out=None):
     """
-    Prepare the data by optionally clipping and copying, and return the array
-    that should be subsequently used for in-place calculations.
+    Prepare the data by optionally clipping and copying, and return the
+    array that should be subsequently used for in-place calculations.
     """
+
     if clip:
         return np.clip(values, 0., 1., out=out)
     else:
@@ -47,38 +49,36 @@ def _prepare(values, out=None, clip=True):
 @six.add_metaclass(InheritDocstrings)
 class BaseStretch(BaseTransform):
     """
-    Base class for the stretch classes, which, when called with an array of
-    values in the range [0:1], return an transformed array of values, also in
-    the range [0:1].
+    Base class for the stretch classes, which, when called with an array
+    of values in the range [0:1], return an transformed array of values,
+    also in the range [0:1].
     """
 
-    def __call__(self, values, out=None, clip=True):
+    def __call__(self, values, clip=True, out=None):
         """
         Transform values using this stretch.
 
         Parameters
         ----------
-        values : `~numpy.ndarray` or list
-            The input values, which should already be normalized to the [0:1]
-            range.
+        values : array-like
+            The input values, which should already be normalized to the
+            [0:1] range.
+        clip : bool, optional
+            If `True` (default), values outside the [0:1] range are
+            clipped to the [0:1] range.
         out : `~numpy.ndarray`, optional
             If specified, the output values will be placed in this array
             (typically used for in-place calculations).
-        clip : bool, optional
-            If `True` (default), values outside the [0:1] range are clipped to
-            the [0:1] range.
 
         Returns
         -------
-        new_values : `~numpy.ndarray`
+        result : `~numpy.ndarray`
             The transformed values.
         """
 
     @property
     def inverse(self):
-        """
-        Return a stretch that performs the inverse operation.
-        """
+        """A stretch object that performs the inverse operation."""
 
 
 class LinearStretch(BaseStretch):
@@ -91,12 +91,12 @@ class LinearStretch(BaseStretch):
         y = x
     """
 
-    def __call__(self, values, out=None, clip=True):
-
-        return _prepare(values, out=out, clip=clip)
+    def __call__(self, values, clip=True, out=None):
+        return _prepare(values, clip=clip, out=out)
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return LinearStretch()
 
 
@@ -110,16 +110,14 @@ class SqrtStretch(BaseStretch):
         y = \sqrt{x}
     """
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         np.sqrt(values, out=values)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return PowerStretch(2)
 
 
@@ -131,22 +129,25 @@ class PowerStretch(BaseStretch):
 
     .. math::
         y = x^a
+
+    Parameters
+    ----------
+    a : float
+        The power index (see the above formula).
     """
 
     def __init__(self, a):
         super(PowerStretch, self).__init__()
         self.power = a
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         np.power(values, self.power, out=values)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return PowerStretch(1. / self.power)
 
 
@@ -158,6 +159,12 @@ class PowerDistStretch(BaseStretch):
 
     .. math::
         y = \frac{a^x - 1}{a - 1}
+
+    Parameters
+    ----------
+    a : float, optional
+        The ``a`` parameter used in the above formula.  Default is 1000.
+        ``a`` cannot be set to 1.
     """
 
     def __init__(self, a=1000.0):
@@ -166,24 +173,34 @@ class PowerDistStretch(BaseStretch):
         super(PowerDistStretch, self).__init__()
         self.exp = a
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         np.power(self.exp, values, out=values)
         np.subtract(values, 1, out=values)
         np.true_divide(values, self.exp - 1.0, out=values)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return InvertedPowerDistStretch(a=self.exp)
 
 
 class InvertedPowerDistStretch(BaseStretch):
     """
-    Inverse transformation for `~astropy.image.scaling.PowerDistStretch`.
+    Inverse transformation for
+    `~astropy.image.scaling.PowerDistStretch`.
+
+    The stretch is given by:
+
+    .. math::
+        y = \frac{\log(y (a-1) + 1)}{\log a}
+
+    Parameters
+    ----------
+    a : float, optional
+        The ``a`` parameter used in the above formula.  Default is 1000.
+        ``a`` cannot be set to 1.
     """
 
     def __init__(self, a=1000.0):
@@ -192,18 +209,16 @@ class InvertedPowerDistStretch(BaseStretch):
         super(InvertedPowerDistStretch, self).__init__()
         self.exp = a
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         np.multiply(values, self.exp - 1.0, out=values)
         np.add(values, 1, out=values)
-        logn(self.exp, values, out=values)
-
+        _logn(self.exp, values, out=values)
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return PowerDistStretch(a=self.exp)
 
 
@@ -222,6 +237,7 @@ class SquaredStretch(PowerStretch):
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return SqrtStretch()
 
 
@@ -233,50 +249,61 @@ class LogStretch(BaseStretch):
 
     .. math::
         y = \frac{\log{(a x + 1)}}{\log{(a + 1)}}.
+
+    Parameters
+    ----------
+    a : float
+        The ``a`` parameter used in the above formula.  Default is 1000.
     """
 
     def __init__(self, a=1000.0):
         super(LogStretch, self).__init__()
         self.exp = a
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         np.multiply(values, self.exp, out=values)
         np.add(values, 1., out=values)
         np.log(values, out=values)
         np.true_divide(values, np.log(self.exp + 1.), out=values)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return InvertedLogStretch(self.exp)
 
 
 class InvertedLogStretch(BaseStretch):
     """
     Inverse transformation for `~astropy.image.scaling.LogStretch`.
+
+    The stretch is given by:
+
+    .. math::
+        y = \frac{e^{y} (a + 1) -1}{a}
+
+    Parameters
+    ----------
+    a : float, optional
+        The ``a`` parameter used in the above formula.  Default is 1000.
     """
 
     def __init__(self, a):
         super(InvertedLogStretch, self).__init__()
         self.exp = a
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         np.multiply(values, np.log(self.exp + 1.), out=values)
         np.exp(values, out=values)
         np.subtract(values, 1., out=values)
         np.true_divide(values, self.exp, out=values)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return LogStretch(self.exp)
 
 
@@ -288,24 +315,31 @@ class AsinhStretch(BaseStretch):
 
     .. math::
         y = \frac{{\rm asinh}(x / a)}{{\rm asinh}(1 / a)}.
+
+    Parameters
+    ----------
+    a : float, optional
+        The ``a`` parameter used in the above formula.  The value of
+        this parameter is where the asinh curve transitions from linear
+        to logarithmic behavior, expressed as a fraction of the
+        normalized image.  Must be in the range between 0 and 1.
+        Default is 0.1
     """
 
     def __init__(self, a=0.1):
         super(AsinhStretch, self).__init__()
         self.a = a
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         np.true_divide(values, self.a, out=values)
         np.arcsinh(values, out=values)
         np.true_divide(values, np.arcsinh(1. / self.a), out=values)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return SinhStretch(a=1. / np.arcsinh(1. / self.a))
 
 
@@ -317,24 +351,27 @@ class SinhStretch(BaseStretch):
 
     .. math::
         y = \frac{{\rm sinh}(x / a)}{{\rm sinh}(1 / a)}
+
+    Parameters
+    ----------
+    a : float, optional
+        The ``a`` parameter used in the above formula.  Default is 1/3.
     """
 
-    def __init__(self, a=1. / 3.):
+    def __init__(self, a=1./3.):
         super(SinhStretch, self).__init__()
         self.a = a
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         np.true_divide(values, self.a, out=values)
         np.sinh(values, out=values)
         np.true_divide(values, np.sinh(1. / self.a), out=values)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return AsinhStretch(a=1. / np.sinh(1. / self.a))
 
 
@@ -344,8 +381,11 @@ class HistEqStretch(BaseStretch):
 
     Parameters
     ----------
-    data : float
-        The data defining the equalization
+    data : array-like
+        The data defining the equalization.
+    values : array-like, optional
+        The input image values, which should already be normalized to
+        the [0:1] range.
     """
 
     def __init__(self, data, values=None):
@@ -362,22 +402,28 @@ class HistEqStretch(BaseStretch):
         else:
             self.values = values
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         values[:] = np.interp(values, self.data, self.values)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return InvertedHistEqStretch(self.data, values=self.values)
 
 
 class InvertedHistEqStretch(BaseStretch):
     """
     Inverse transformation for `~astropy.image.scaling.HistEqStretch`.
+
+    Parameters
+    ----------
+    data : array-like
+        The data defining the equalization.
+    values : array-like, optional
+        The input image values, which should already be normalized to
+        the [0:1] range.
     """
 
     def __init__(self, data, values=None):
@@ -387,16 +433,14 @@ class InvertedHistEqStretch(BaseStretch):
         else:
             self.values = values
 
-    def __call__(self, values, out=None, clip=True):
-
-        values = _prepare(values, out=out, clip=clip)
-
+    def __call__(self, values, clip=True, out=None):
+        values = _prepare(values, clip=clip, out=out)
         values[:] = np.interp(values, self.values, self.data)
-
         return values
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return HistEqStretch(self.data, values=self.values)
 
 
@@ -410,6 +454,14 @@ class ContrastBiasStretch(BaseStretch):
         y = (x - {\\rm bias}) * {\\rm contrast} + 0.5
 
     and the output values are clipped to the [0:1] range.
+
+    Parameters
+    ----------
+    contrast : float
+        The contrast parameter (see the above formula).
+
+    bias : float
+        The bias parameter (see the above formula).
     """
 
     def __init__(self, contrast, bias):
@@ -417,11 +469,10 @@ class ContrastBiasStretch(BaseStretch):
         self.contrast = contrast
         self.bias = bias
 
-    def __call__(self, values, out=None, clip=True):
-
-        # As a special case here, we only clip *after* the transformation since
-        # it does not map [0:1] to [0:1]
-        values = _prepare(values, out=out, clip=False)
+    def __call__(self, values, clip=True, out=None):
+        # As a special case here, we only clip *after* the
+        # transformation since it does not map [0:1] to [0:1]
+        values = _prepare(values, clip=False, out=out)
 
         np.subtract(values, self.bias, out=values)
         np.multiply(values, self.contrast, out=values)
@@ -434,12 +485,23 @@ class ContrastBiasStretch(BaseStretch):
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return InvertedContrastBiasStretch(self.contrast, self.bias)
 
 
 class InvertedContrastBiasStretch(BaseStretch):
     """
     Inverse transformation for ContrastBiasStretch.
+
+    Parameters
+    ----------
+    contrast : float
+        The contrast parameter (see
+        `~astropy.visualization.ConstrastBiasStretch).
+
+    bias : float
+        The bias parameter (see
+        `~astropy.visualization.ConstrastBiasStretch).
     """
 
     def __init__(self, contrast, bias):
@@ -447,13 +509,10 @@ class InvertedContrastBiasStretch(BaseStretch):
         self.contrast = contrast
         self.bias = bias
 
-    def __call__(self, values, out=None, clip=True):
-
-        # As a special case here, we only clip *after* the transformation since
-        # it does not map [0:1] to [0:1]
-
-        values = _prepare(values, out=out, clip=False)
-
+    def __call__(self, values, clip=True, out=None):
+        # As a special case here, we only clip *after* the
+        # transformation since it does not map [0:1] to [0:1]
+        values = _prepare(values, clip=False, out=out)
         np.subtract(values, 0.5, out=values)
         np.true_divide(values, self.contrast, out=values)
         np.add(values, self.bias, out=values)
@@ -465,4 +524,5 @@ class InvertedContrastBiasStretch(BaseStretch):
 
     @property
     def inverse(self):
+        """A stretch object that performs the inverse operation."""
         return ContrastBiasStretch(self.contrast, self.bias)