+++ astropy/visualization/ui.py
@@ -4,11 +4,16 @@ import numpy as np
 
 from .interval import (PercentileInterval, AsymmetricPercentileInterval,
                        ManualInterval, MinMaxInterval)
-
 from .stretch import (LinearStretch, SqrtStretch, PowerStretch, LogStretch,
                       AsinhStretch)
+from ..utils.decorators import deprecated
 
 
+__all__ = ['scale_image']
+
+
+@deprecated('1.3',
+            alternative='`~astropy.visualization.mpl_normalize.simple_norm`')
 def scale_image(image, scale='linear', power=1.0, asinh_a=0.1, min_cut=None,
                 max_cut=None, min_percent=None, max_percent=None,
                 percent=None, clip=True):